+++ CHANGELOG.md
@@ -23,7 +23,10 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 ## [Unreleased]
 
+## [1.5.0] - 2023-07-10
+
 *   Add `tf.nest`/`tf.data` compatibility for `DataclassArray`.
+*   Fix import error due to `etils`
 
 ## [1.4.2] - 2023-07-10
 
@@ -72,7 +75,8 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 
 *   Initial release
 
-[Unreleased]: https://github.com/google-research/dataclass_array/compare/v1.4.2...HEAD
+[Unreleased]: https://github.com/google-research/dataclass_array/compare/v1.5.0...HEAD
+[1.4.2]: https://github.com/google-research/dataclass_array/compare/v1.4.2...v1.5.0
 [1.4.2]: https://github.com/google-research/dataclass_array/compare/v1.4.1...v1.4.2
 [1.4.1]: https://github.com/google-research/dataclass_array/compare/v1.4.0...v1.4.1
 [1.4.0]: https://github.com/google-research/dataclass_array/compare/v1.3.0...v1.4.0